+++ package.json
@@ -5,6 +5,7 @@
     "dotenv": "^16.4.5",
     "nodemon": "^3.1.7",
     "prettier": "^3.3.3",
+    "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.6.3",
